Use explicit uint32_t casts in day subtraction arithmetic
Ensure each operand is promoted to uint32_t before subtraction
to prevent any possibility of unsigned underflow, even if
sanitization checks were somehow bypassed.

+++ Core/Src/main.c
@@ -2240,7 +2240,7 @@ static void calendar_add_minutes(rtc_calendar_t *cal, uint32_t minutes_to_add)
         }
         else
         {
-            days_to_add -= (uint32_t)(dim - cal->day + 1u);
+            days_to_add -= ((uint32_t)dim - (uint32_t)cal->day + 1u);
             cal->day = 1;
             cal->month++;
             if (cal->month > 12)
